requirements.txt
streamlit==1.28.0
opencv-python==4.8.1.78
numpy==1.26.0
Pillow==10.1.0
matplotlib==3.8.0